more deploy WIP

+++ deploy_to_github_pages.sh
@@ -3,34 +3,44 @@ set -o errexit
 set -o pipefail
 set -o nounset
 
-# Check out gh-pages branch
-git checkout gh-pages
+echo "Deploy to github pages..."
+read -p "Are you sure? This will wipe any uncommited local changes" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    # Check out gh-pages branch
+    git checkout gh-pages
 
-# Pull master into it
-git reset --hard master
+    # TODO: if error, give warning that probably need to create the branch (i.e. git checkout -b gh-pages)
 
-# Compile
-npm install
-npm run compile
+    # Pull master into it
+    git reset --hard master
 
-# Move app to /tmp
-# TODO:
-rm -rf /tmp/app
-mv app /tmp/
+    # Compile
+    npm install
+    npm run compile
 
-# Then delete everything else (including this script)
-rm -rf ./*
+    # Move app to /tmp
+    # TODO:
+    rm -rf /tmp/app
+    mv app /tmp/
 
-# Then move contents of app to top level
-mv /tmp/app/* ./
+    # Then delete everything else (including this script)
+    rm -rf ./*
 
-# Commit and force push
-git add *
-git commit -a -m 'update'
-git push -f
+    # Then move contents of app to top level
+    mv /tmp/app/* ./
 
-# Switch back to master
-git checkout master
+    # Commit and force push
+    git add *
+    git commit -a -m 'update'
+    git push --set-upstream origin gh-pages
 
-# Site code should now be at https://github.com/<user>/<project>/tree/gh-pages
-# And should be live at https://<user>github.com/<project>
+    # Switch back to master
+    git checkout master
+    npm install
+
+    echo "Site should be live at https://<user>/github.com/<project>"
+    echo "And code should now be at https://github.com/<user>/<project>/tree/gh-pages"
+
+fi
